feat: restructure folder
