+++ src/common/multer/multer.service.ts
@@ -6,7 +6,7 @@ export class MulterService {
   constructor() {}
 
   saveFile(file: Express.Multer.File): string {
-    join(__dirname, '../../uploads', file.filename);
+    join(__dirname, '../../../uploads', file.filename);
     return file.filename;
   }
 }
