+++ src/contact/contact.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ContactController } from './contact.controller';
 import { ContactService } from './contact.service';
-import { MulterService } from '../common/multer.service';
+import { MulterService } from '../common/multer/multer.service';
 
 @Module({
   controllers: [ContactController],
