+++ src/main.ts
@@ -8,7 +8,7 @@ import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 async function bootstrap() {
   const app = await NestFactory.create<NestExpressApplication>(AppModule);
 
-  app.useStaticAssets(join(__dirname, '../uploads'), {
+  app.useStaticAssets(join(__dirname, './../uploads'), {
     prefix: '/uploads',
   });
