+++ src/user/user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
-import { MailService } from '../common/mail.service';
+import { MailService } from '../common/mail/mail.service';
 
 @Module({
   providers: [UserService, MailService],
