+++ test/test.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TestService } from './test.service';
-import { HashidService } from '../src/common/hashid.service';
+import { HashidService } from '../src/common/helper/hashid.service';
 
 @Module({
   imports: [],
